requirements.txt
absl-py==0.7.1
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
Click==7.0
cloudpickle==1.4.1
cycler==0.10.0
Cython==0.29.15
decorator==4.4.0
deepdish==0.3.6
defusedxml==0.6.0
dominate==2.4.0
dotmap==1.3.17
entrypoints==0.3
Flask==1.1.1
gast==0.2.2
grpcio==1.16.1
h5py==2.9.0
imagecodecs-lite==2019.12.3
imageio==2.9.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.14.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras-tqdm==2.0.1
kiwisolver==1.1.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
metayaml==0.25
mistune==0.8.4
mmcv==0.5.6
mock==2.0.0
mpmath==1.1.0
msgpack==0.6.1
msgpack-numpy==0.4.4.2
nbconvert==5.5.0
nbformat==4.4.0
nose==1.3.7
notebook==5.7.8
numexpr==2.7.1
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plotly==4.8.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.12.2
ptyprocess==0.6.0
Pygments==2.4.0
pyparsing==2.4.0
pyrsistent==0.15.2
PySide2==5.14.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.3
pyzmq==18.0.1
qtconsole==4.4.4
remi==2020.2.5
retrying==1.3.3
scikit-learn==0.21.3
scipy==1.2.1
scipy-stack==0.0.5
seaborn==0.9.0
Send2Trash==1.5.0
shiboken2==5.14.1
six==1.12.0
sk-video==1.1.10
sympy==1.4
tables==3.6.1
tabulate==0.8.3
tensorboard==1.13.1
tensorboardX==2.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-probability==0.6.0
tensorpack==0.9.4
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tifffile==2019.7.26.2
torch==1.4.0
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
